requirements.txt
alembic==0.9.1
Babel==2.4.0
beautifulsoup4==4.5.3
click==6.7
Flask==0.12.1
Flask-Assets==0.12
Flask-Babel==0.11.2
Flask-Script==2.0.5
future==0.16.0
gunicorn==19.7.1
httplib2==0.10.3
itsdangerous==0.24
Jinja2==2.9.6
line-bot-sdk==1.1.0
Mako==1.0.6
MarkupSafe==1.0
oauth2==1.9.0.post1
python-editor==1.0.3
pytz==2017.2
requests==2.13.0
SQLAlchemy==1.1.9
virtualenv==15.1.0
virtualenvwrapper-win==1.2.1
webassets==0.12.1
Werkzeug==0.12.1
